rename DeleteBox file name
